Update Process_GPS.R
This change should make the check for the duplicates faster. It uses a loop to subset the full dataframe into chunck for only 1 gps_id, and looks for duplicates only for that logger, instead of looking in the whole database. This seems more effective for ECOTONE users that may have the loggers running for a month or more.